+++ CHANGELOG.md
@@ -18,7 +18,8 @@ All notable changes for each version of this project will be documented in this 
     - **Breaking Change** The `onSlideChanged`, `onSlideAdded`, `onSlideRemoved`, `onCarouselPaused` and `onCarouselPlaying` outputs have been renamed to `slideChanged`, `slideAdded`, `slideRemoved`, `carouselPaused` and `carouselPlaying` to not violate the no on-prefixed outputs convention. Automatic migrations are available and will be applied on `ng update`.
 - `IgxRadio`, `IgxRadioGroup`
     - Added component validation along with styles for invalid state
-
+- `IgxAvatar`
+    - **Breaking Change** The `roundShape` property has been replaced with `shape` property that can be either `square` or `rounded`. The default shape of the avatar from now on is rounded. 
 
 ## 15.0.1